// src/messages/enums/BackendRoutesEnum.ts
export enum BackendRoutesEnum {
    API_URL = "/api/v01",
    HEALTH_CHECK = "/health-check",
    USERS = "/users",
    ROLE = "/roles",
    CONTENT = "/content",
    ARTICLES = "/articles",
    FILES = "/files",
    CATEGORY = "/category",
    PAGE = "/page",
    FILE = "/file",
    TAG = "/tag",
    AUTH = "/auth",
    LOCAL_AUTH = "/local",
    FACEBOOK_AUTH = "/facebook",
    TWITTER_AUTH = "/twitter",
    GOOGLE_OAUTH = "/google/oauth",
    GOOGLE_OAUTH2 = "/google/oauth2",
    OPTIONS = "/options"
}